RaspberryPiVersion/requirements_rpiversion.txt
python3-pandas
python-dateutil
python3-numpy
python3-flask
python3-pigpio
